final changes to end this saga
